set time with NTP when a network interface is brought up
DNS queries about pool.ntp.org are made using the DHCP-provided nameservers,
when possible, in the script that sets the clock:
/etc/NetworkManager/dispatcher.d/50-ntp.sh.

Else, we use OpenDNS as Incognito does.

However, we really need /etc/resolv.conf to be setup in a way that makes the
rest of the system use the local (torified) DNS cache. This is achieved thanks
to the − properly configured − resolvconf package. Our custom dhclient
configuration can then be dropped: it was only useful to supersede
DHCP-provided nameservers, which we do not want to do anymore since we need this
information, as explained above.

As live-helper manages /etc/resolv.conf in a way that makes it too hard for us
to ensure the image ends up with the needed symlink to
/etc/resolvconf/run/resolv.conf, this symlink mangling is done at boot time,
thanks to a live-initramfs script:
config/chroot_local-includes/usr/share/initramfs-tools/scripts/live-bottom/00resolv_conf