Design draft: get a removed TODO item back.
Vaguely stating that "Efforts should also be made to avoid browser fingerprint"
in the specification is ok, but we also need to document how this is actually
implemented in T(A)ILS, what compromises are made, and what the actual results
are. See the recent thread about this on the (private) dev mailing list.