Use normal internal links rather than the tails_website shortcut.
Let's not start introducing broken links in static builds of this wiki (such as
the one included in birary images we ship) where we can avoid it.

This shortcut was solely added so that we can refer to the T(A)ILS website as
the canonical location. It should only be used when really needed, not as a
general purpose replacement of usual ikiwiki's internal linking syntax.