Disable init system's concurrency at the end of the boot sequence.
Dependency based boot sequencing is now the default in Debian.

We want to benefit from it to decrease boot time,
but we need to disable it for the shutdown sequence in order to have our memory
wiping hooks working in a reliable way... or at least to help fixing their bugs.