Move information about latest release to tiny files that are included where needed.
The same effect was previously attempted using [[!template ...]] which proved to
be a mess to deal with for HTML and translatable pages.

Note that [[!inline ...]] prepends a newline before the inlined content, which
makes it impossible to use two inline directives in the same URL, hence the
seemingly unnecessary information duplication. At least that information now is
gathered in a central place and not exposed to translators unless needed (i.e.
only the release date is translatable).