pdnsd: query ttdnsd for anything but .onion/.exit.
The Tor resolver is still used for .onion/.exit.

It would be nice to query first the Tor resolver, and to fallback to ttdnsd in
case the former is not able to answer our query. However, the Tor resolver sends
an empty reply rather than an error when it does not support the type of the
query (e.g. MX), so we're forced to use ttdnsd by default instead.