Pass "-t 1" to htpdate that converts it to wget --dns-timeout 1.
When we run it from htpdate, wget shouldn't attempt to resolve DNS itself,
by asking to pdnsd (as it does), because the hostname it needs is in /etc/hosts.
But for whatever reason, it tries that anyway, so it'd better not try too long
'cause it won't work, and as long as it waits, the user waits and has no Tor nor
Vidalia. This is a blocking operation, y'know.