Make kexec find out what kernel was booted, and load this one.
We'll soon ship several kernels for PAE / NX bit support: 486, 686-pae, amd64.

Therefore, if the system was booted on a 64-bit kernel, plenty of memory may
have been used, that need to be wiped on shutdown; in such a situation, the
memory wipe kexec'd kernel must be 64-bit too: a 32-bit kernel may be unable to
access all memory that may have been used during the runtime of Tails.