Torify wget.
We set the proxy envvars in /etc/environment, but they are not read by
init scripts, for instance. In particular this makes the the htpdate
init script fail when htpdate calls wget, since no proxy is set for
it. Setting proxies in wgetrc may be more reliable than trusting the
envvars in other situations as well.