Add Alt+Shift shortcut to switch keyboard layout.
This is the default layout switching shortcut used in Windows, which
probably is what people are used to. Note that this causes a layout
switch to occur for other keyboard shortcuts using Alt and Shift,
which is annoying. In Tails this currently only accects the
Shift+Ctrl+Alt+<arrows> shortcuts for moving the selected window to
another virtual deskop, so this is IMHO acceptible.