iceweasel: enable TLS false start.
TBB does this since last December.

There are a few web servers that don't support it, that's why it is disabled by
default in Firefox (https://bugzilla.mozilla.org/show_bug.cgi?id=658222), but
well, their number seem to be small enough to be ignored. The advantages
(including not diverging from TBB) are probably worth it.