Do not display storage volumes on Desktop.
... by disabling /apps/nautilus/desktop/volumes_visible GConf entry.

Enabling that GConf setting avoids displaying the bind-mounted persistent
directories on the Desktop, and reduces user confusion.

It also is a first step towards a bigger UI change: GNOME3 does not manage the
Desktop anymore, so volume icons and other Desktop icons are meant to disappear
anyway. It implies we'll have to move all Desktop icons elsewhere. Let's start
this move now: this will smooth the UI change Wheezy will carry for our users,
by applying some of it progressively.