Remove obsolete noswap boot parameter.
live-boot has been respectful of the "swap is opt-in" rule since the last time
we fixed it there, that is almost 3 years ago. There is no parameter like noswap
to *disable* swap anymore, but only the swapon parameter that allows to enable
it. Therefore, it looks like we can safely simplify our configuration a bit, and
stop carrying such obsolete stuff.