vagrant: Add suport for HTTP proxy build settings
We support three options: external proxy, a proxy inside the virtual machine
and no proxy at all.

The first and the last are both trivial to implement.

For the in-VM proxy, we install and configure apt-cacher-ng during
provisioning. We do that all the time to support different proxy options once
the VM is started.

The custom apt-cacher-ng configuration is required to support downloads of `.o`
and `.tar.bz2` files by the firmware installers.