Prevent memlockd unload on shutdown
memlockd keeps our tails-kexec file in memory. This is crucial to have
emergency shutdown on media removal. So in order to prevent any race
condition regarding the avaibility of the files needed to start the
memory wiping kernel, we do everything we can to prevent memlockd from
being unloaded during shutdown. Namely:

 1. We properly remove stop symlinks to the memlockd initscript using
    `update-rc.d`.
 2. We add memlockd PID to the process that are ommited by the sendsigs
    script (responsible for sending TERM and KILL signals to remaining
    processes).
 3. Do not require that memlockd be stopped before stopping
    tails-sdmem-on-media-removal.

In addition, we fix initscript ordering related to
tails-sdmem-on-media-removal.

These changes towards memlockd are too specific to Tails' needs to be
upstreamed: if memlockd is used to keep files in memory from rw media,
having memlockd running would prevent proper unmounting. That would be
bad.