Make sure USB boot devices are owned by the 'disk' group.
Tails is hit by Debian bug #645466: udev sets group ownership of USB
boot devices to 'floppy', which the live user is a member of, and
hence it can write to the boot device. To fix this we make sure that
the boot device (and any other partitions on it) are owned by 'disk'.

This fixes bugs/writable_system_disk:_belongs_to_floppy_group