Be specific in which linux-image packages has to be installed
linux-latest (which among others build `linux-image-2.6-486`) is sometimes not
up-to-date wrt. the kernel packages that are currently in unstable. In order
to avoid troubles, let's install a specific version instead of the default one.

This does not really grow the maintaince burden as apt_preferences already
needed to be updated to have the latest linux package names.