Replace sdmem by custom program using hugetlb for memory wipe
The way we previously wiped memory when exiting Tails was to start many
instances of sdmem in parallel and let them be killed by the kernel
when it was reaching OOM condition.

This gives us little control over what the processes are currently doing, on
how much memory they are effectively going to wipe and the current state of
progress.

So, instead of relying on sdmem, we now use a small program written in C with
the following design:

Memory wipe is done on memory areas allocated using mmap(). When the
memory is to be filled with zeros, it relies on the combination of
MAP_ANONYMOUS | MAP_POPULATE | MAP_SHARED flags which makes the kernel
clear the memory. Otherwise, it uses memset() on the allocated area.

First, pages are allocated using MAP_HUGETLB. Once all huge pages are
allocated, we continue the process using "usual" page size until we have
taken care of the amount of memory given on the command line.

A new process is spawned after one process has wiped 1 GB in order to
overcome the limit of maximum memory addressable by a single process
on 32-bit architectures.

Using huge pages should result in a faster wipe, as less page allocations are
required to clean up the whole memory. See
<http://linuxgazette.net/155/krishnakumar.html> for an introduction on huge
pages and hugetlb.

The program uses command-line arguments to know how much memory still needs to
be wiped, allowing to display a nice progress bar.

This C helper is still invoked using a initramfs script, with the following
difference from the previous 'sdmem' script:

 * Use vm.overcommit_memory=2 and vm.overcommit_ratio=100: the C program
   works better when it gets a nice ENOMEM when requesting memory instead
   of being killed by the kernel.
 * Set vm.lowmem_reserve_ratio to values that should not preserve much
   of the low memory areas (DMA, DMA32). This is the worst thing to do
   on a system with running drivers, but at this stage we should not
   have much loaded, and this allows to actually wipe more memory.
 * Runs in init-top and not init-premount: there is no need to have udev
   working to perform the wipe. This saves a few seconds.

The script also takes care of mounting the 'hugetlbfs' pseudo-filesystem that
is needed to use MAP_HUGETLB.

And finally, it computes the amount of memory that needs to be cleaned by the C
helper. It purposely substract the amount of memory in the "cached" state.
These bytes have already been overwritten by loading files, so the previous content
is already gone. This helps us to stay away from the OOM-condition and reach
the end of our nifty progress bar.