Do not assume page sizes in hugetlb_mem_wipe
We previously hardcoded page sizes in hugetlb_mem_wipe. This is problematic as,
at least, size of huge pages is 2 MB on x86 PAE and 4 MB on x86 without PAE.

We now use the initramfs script to retrieve appropriate value from the system
settings (through `getconf` and `/proc/meminfo`) and give them to the
hugetlb_mem_wipe command line.