Use 127.0.0.1 instead of localhost to access I2P router console
Now that Iceweasel has Torbrowser patches, the `socks_remote_dns` option
enabled by Torbutton will prevent any name resolving that is not going through
a proxy. So we need to change the bookmark for I2P router console to directly
use the IPv4 address instead of 'localhost'.

Given that change, there is no need to keep the rules for 'localhost' in
FoxyProxy, let's just keep only the one that applies to '127.0.0.1'.