Let GTK handle menu positioning.
Our custom menu positioning functions do not add anything of value,
and in fact make the menus unusable when the status icon is in a panel
located on the bottom of the screen. That made gpgApplet unusable when
Windows XP camouflage was activated. GTK seems to have logic for
handling this so dropping the functions instead of making them handle
this case seems like the ideal solution.