Create wrappers for (re)starting Tor and Vidalia.
If Vidalia is running, and Tor is restarted, then we also want to
restart Vidalia. This is because Vidalia doesn't re-connect to Tor
automatically, so the user has to restart it to be able to control Tor
again. Also, any options set by Vidalia will be lost since they
weren't written to torrc, which causes Tor to reach an inactive state
if it's restarted in "bridge mode".