Add "Shutdown Helper" Gnome applet.
This applets implements todo/better_power_off_button. It's just a
simple button that opens a menu with options to lock the screen, halt
or reboot. The idea is to replace the current emergency shutdown
button (which lacks any kind of action verification) that's a bit too
easy to miss-click when closing a maximized window, due to the close
proximity of the X-button.