Remove pinning for xul-ext-adblock-plus.
Now that we've added this package to our APT repository (testing, devel and
experimental suites), our pinning fallback configuration would anyway fetch it
from there, and not from sid anymore. So, the pinning is not useful anymore.

We'll soon drop the package from our APT repository and add some pinning back
and fetch the package from squeeze-backports, but we're not there yet.