Make parallel sdmem faster.
We avoid running too many parallel instances (which adds overhead) by
only running one instance per 2048 MiB of RAM (which is below the x86
per-process limitation). This also seems to play better with
VirtualBox, which often hangs when running too many sdmem in parallel.

Furthermore we kill all sdmem instances as soon as one of them has
been oom killed; after all, that implies that the memory has been
completely filled, so we're done. If we don't kill the others they may
continue running unecessarily, filling the freed memory of the killed
instance, and so on.

Another consequence of this is cleaner output so that the debug
message (that memory dumping can ensue) will be easier to spot when
running with boot option debug=wipemem. That will be a helpful cue
for the automated test suite.