Don't add "quiet" to the kernel command-line ourselves.
Else, it appears twice as live-build's lb_binary_syslinux adds it too.

Historically, we've been adding it ourselves on top of that because some other
live-build's bootloader config handling scripts (such as lb_binary_yaboot) do
not add it, but since we gave up the PowerPC support attempt, we're now only
interested in syslinux, so let's make it easier for the general case, e.g.
when one wants to remove the "quiet" parameter as suggested by our faulty
startup debugging documentation.