Workaround build failure in lb_source, after creating the ISO.
At `lb source` time, when `lb chroot_sources install` wants to run `apt-get
upgrade`, it fails the build with `The value 'wheezy' is invalid for
APT::Default-Release as such a release is not available in the sources` (due to
the APT proxy not being configured at this time, the last `apt-get update`
failed and the package indices are empty). If one simply removes
`APT::Default-Release`, then the pinning is broken when indices are
not available.