Fetch a few packages from Wheezy instead of unstable.
Namely: spice-vdagent, libregexp-common-perl, macchanger, service-wrapper,
libservice-wrapper-java and libservice-wrapper-jni.

Wheezy has the versions we've been installing for a while,
so let's avoid having unstable push a newer one to us uselessly at some point.

Note that at the time of my writing, the versions in sid and in Wheezy are the
same, so this commit is effectively a no-op as of today: it is merely
a safeguard for the future.