Don't test safe access rights in old releases.
We only care if the current release has it. If we check for it in the
scenario setting up the "old" Tails USB installation using a version
without the fix, e.g. versions <= 0.18, it will fail, resulting in
that some of the subsequent scenarios will fail for no good reason.