Rewrite Unsafe Browser branding patching mechanism for FF24.
Mozilla's .ja and .jar files are now optimized in a way that makes it basically
non-standard ZIP archives. 7z can't read them, so let's switch to unzip, that
warns and cries but manages to handle it (still have to ignore its non-zero exit
code, but well).

So, we need to install unzip.

https://bugzilla.mozilla.org/show_bug.cgi?id=605524
https://bugzilla.mozilla.org/show_bug.cgi?id=595473