Disable Torbutton 1.6's check for Tor.
First, this check did not exist in Torbutton << 1.6, so this feature was never
used in Tails yet.

Second, we cannot use the local Tor check with the control port, so we have
been relying on the remote Tor check since feature/ff24 has imported Torbutton
1.6.x.  Unfortunately, the new check.torproject.org breaks the remote Tor
check, and nobody but us cares^Wdiscovered it yet, as the only currently
supported usecase of Torbutton is the TBB, which itself uses the local check.

So, the shortest and sanest path to fixing the "about:tor wrongly pretends that
Tor is not configured properly, because the remote Tor check is broken" seems
to simply disable this check. If someone strongly wants Tails to have such a
Tor check feature, they're welcome to fix the new backend on
check.torproject.org, and then we can revert this commit.

Note that this Iceweasel pref was introduced with a Tails-specific patch in our
Torbutton 1.6.4.1-1 package. A pull request will be submitted shortly so that
this feature is upstream'd.