Add fail-safe if MAC spoofing fails.
To prevent leaking the real MAC address upon MAC spoofing failure we
now try to unload the module before starting NM if such a failure is
detected. If the module can't be unloaded, we simply don't start
NM. In both cases an informative warning notification is shown.

This commit depends on a TG with commit a4451b3, "Don't start NM at
post-login". In order to get better control of whether to start NM or
not (depending on the error) we instead do it in this script, which is
called from the same place in TG any way.