Get back to installing Linux 3.10-1.
This reverts commits 0c844df1fb9b9b44ddf92c1150ed40135c651f60
and 9cd510515030f091b4ab555e82f8b8114b2a5c44.

The APT pinning for linux*3.10* is actually useless, as these packages are not
in Debian testing anymore and will be covered by the general pinning rules that
apply to our own APT repository. Let's keep these lines around so that we don't
forget to adapt them once we move forward to a newer kernel again.