Use the keyboard instead of clicking around to enable persistence.
I have seen some weird race condition cases where the "No" persistence button is
still clicked, and the mouse cursor is over the "Yes" button.

Might it be that my test system is too fast and the Greeter displays widgets
before they are really ready to be clicked?

I have tried to wait in various ways for the "Yes" button to be clicked, and
after a few hours of unsuccessful attempts, I'm giving up and going with the
easy keyboard -based approach, that is simple and worked on the first try.