Make it possible to manually start IBus.
By always exporting the environment variables it is
possible for the user to manually start IBus in any
locale using "System->Preferences->IBus Preferences".
IBus need to actually be started for the french
keyboard shortcuts bug to appear, so this is safe to do.

Also update the description for current behaviour.