Change MAC spoof verification method.
Some hardware fails to report the permanent MAC address for
`macchanger -p`. To deal with it we simply record the MAC address
before and after trying to spoof it, and compare them. However, since
macchanger truly picks a random value, it can randomly pick the real
MAC address so there's a 1/2^24 probability that nothing really fails
but the fail safe will trigger any way.