Start Tor Launcher and time syncing concurrently.
Having Tor Launcher block will leave it and Tor in a deadlock if the
system clock is way off. The time syncing stuff has to run in parallel
to deal with that.

Unfortunately, in some situations the time syncing script may restart
Tor, which will make Tor Launcher lose its control port connection,
leaving it with a dead progress bar. These are:

1. We failed to get a consensus. In this situation nothing works well
   any way, so whatever we do isn't much of a regression.
2. We get a consensus, but the system time is outside its interval.
   In this case we're guaranteed to make Tor work.

The best we can do at the moment seems to be to to simply kill Tor
Launcher, and ensure that it won't restart upon reconnect. At least
that deals with 2 in pretty nice manner, while 1 isn't any worse off
really.