Drop the obsolete --tasks option.
This has been "replaced" by the fancy "!" prefix and "Packages" script, that
calls grep-aptavail. We would have wanted to use it to install all packages with
Priority=standard. The problem is that this thing of course also returns
standard priority packages that are only available in sid, and that we don't
want to install.

We'll therefore have to manually install the "standard" packages we need.