Move cpu and mem checks to the `build` task.
Previously, when they were checked in `vm:up` *only* when issued while
the VM already is up, these checks weren't run if one issues a `build`
when the VM is off. Now we'll fail earlier with a more informative
error message, and it looks like a more logical home for them too.