Configure the keyboard model and layout used in the GNOME session, accordingly to what the user chose in the Greeter.
This depends on Tails Greeter to save the relevant environment variables
to /var/lib/tails-user-session/keyboard.

Note that we don't set layouts to [chosen one, US] anymore, as I could not make
it work properly: in this case, regardless of the order in which we set it, the
US layout wins and is applied by default in the session. Anyway, it's easy
enough to either directly choose US in the greeter (when one wants the GUI in
their preferred language, and a US keyboard layout), or to add the US layout in
the GNOME settings (when using both layouts in the same session). So, this seems
like an acceptable regression to me.