Fix the image used for detecting Iceweasel.
In commit 13dce7f it was changed from the title to the task icon, and
since we wait_and_click it, that causes a minimize + maximize that may
result in key presses getting lost.

While we're at it we also rename it to *Window.png, to be consistent
with how we have named the other similar images, and to hopefully
prevent a similar bad update in the future.