Create the tails-persistence-setup user with the same UID/GID it had on Tails/Squeeze. (Closes: #7343)
Else, our various checks for safe access rights on persistence.conf fail,
as the ACL stores numerical UID values, and the saved UID does not match
the one tails-persistence-setup gets in Tails 1.1~beta1.

Also remove the --quiet flag: it makes adduser silently ignore failures in
case a user with the same UID already exists.

Also rename the hook to create the tails-persistence-setup user before others,
so that its UID is still free.

Note that we cannot use the mutually exclusive --gid and --group when calling
adduser. That's why we create the group first, with the desired GID.