Modify a fresh Iceweasel profile instead of the system one.
If we're in camouflage mode we want it to be applied to the Unsafe
Browser as well. However, the camouflage activation script modifies
the *user's* own profile, not the system one, so it must be present at
this point, and then it's easier to also make all the other changes to
the user's profile since any changes to the system profile will not be
copied back in.