Start I2P via NetworkManager hook
Since the user clearly is interested in starting I2P, it shouldn't be
necessary to manually start I2P anymore. Therefore the old I2P menu
entry has been changed to start the I2P Browser and NetworkManager
handles the starting of I2P *if* 'i2p' is added by the user to the boot
prompt.