Install TBB in a custom way instead of installing Iceweasel.
Instead of installing the iceweasel package (and friends) we now fetch
the TBB tarballs at build-time and extract the Tor Browser, browser
add-ons and langpacks, and browser user data into sensible places. We
try to keep something like the old organization as much as possible.

Unfortunately this drops language support to whatever subset that the
TBB supports.