Update ISO verification screenshots with the truncated ones (Closes: #7259)
This not only happens in Tails, and will probably stay like this for a while.
But still, they needed to be updated because on GNOME 3 there are no icons
anymore and so the old ones look quite different. Having the truncated
notifications in the screenshots is actually the most accurate thing that we can
do here.