Fix automated test suite for migration to TBB.
Beyond s/Iceweasel/Tor Browser/ (incl. renaming affected images) we
also try to rely more on clicking on GUI elements instead of keyboard
shortcuts. The TBB is noticeably slower on start, and often loses key
presses at that stage, so clicking is much more reliable then.

The most remarkable change may be the switch to the step "the Tor
Browser has started and loaded the startup page". It seems a fair
amount of complexity is required to reliably stop the loading of the
startup page so this simpler approach seems better, and as a bonus it
actualy tests the startup page. The issue described in the removed
comment doesn't seem very relevant any more, too.