Revert "$HOME(pidgin): hide formatting toolbar" (Closes: #7356)
This reverts commit c4a2668bdf49d803e32e3cfc94825115461a9d50.

Having the formatting toolbar displayed in Pidgin is the original default, and
it makes the OTR status more explicit by displaying it with words.

Without this toolbar, the OTR status have to be inferred either from the chat
log, or from the small logo on the right of the OTR menu.