Make the Turbutton button appear in the Tor Browser.
Setting extensions.torbutton.inserted_button to true hides the butten
and is recommended by TBB 4.x's start-tor-browser script when using a
system-wide Tor instance. Probably that's bacause normally an
appropriate environment isn't set up to make e.g. Torbutton's new
identity feature work, so hiding it is a good thing, but in Tails we
do set all that up and want to keep the button.