We should not ignore missing persistent source directories.
That should imply that something has gone very wrong. I'm not sure why
we've been ignoring them (certainly it would deserve a comment!) but I
can imagine there was some thought about the "old iso with fewer
presets than the current iso" situation that was just solved, but that
wouldn't be enough for some of the other checks.