Slightly simplify code, and remove magic number.
This is "slightly" more straigh-forward, and not having to bump the
number is nice, especially now when the list of presets is dynamically
fetched from the Tails system being tested. From now on there should
be no need to add support for new persistence presets in the test
suite without breaking it, and they will benefit from the very basic
and general testing we do of each of them.