Remove test.
The Unsafe Browser shutdown is now so fast that this test sometimes
fails (because the teardown has completed before we start the Unsafe
Browser again).

While we could do artificial stuff (like `kill -9`:ing the script) to
prevent the controlled teardown from happening, the point is lost when
it then pretty much degrades into (and always have been very similar
to) the "Starting a second instance of the Unsafe Browser results in
an error message being shown" scenario.