Test 'Tor is ready' notification in a separate scenario (Will-fix: #8714).
Sikuli isn't reliably detecting the notification, either because it
disappears too quickly given Wheezy's GNOME3's hardcoded timeout of
five seconds, or because it fails to show at all. Any way, this often
introduces failure of the 'Tor is ready' step, so this should improve
the robustness of the test suite.