Create the environment for remote shell calls more realisitcally.
We do an actual login (via `su`) and extract the resulting environment
so that e.g. LANG is set and /etc/environment gets sourced which sets
PATH so that we look for executables in /usr/local/*bin, which truly
matters since some of them are wrappers around executables found in
some other PATH-dir.