Set network state according to the background when restoring from snapshots.
This makes sense for keeping the state (of the network's "plugged"
state) after restoring from a snapshot consistent with the one before,
especially since new VM instances always start with the network
plugged -- if "the network is unplugged" is in the background, the
network will suddenly be plugged after resotring from the background
snapshot.

In fact after the VMNet split (out of VM) seems to expose this issue:
when restoring from a snapshot where the network is unplugged, and
then plugging the network, I get 1 minute delay and errors like
"device not ready (errno=-16), forcing hardreset" in dmesg. That these
strange things didn't happen before the VMNet split is probably just a
coincidence, some how.