Use @check_tor_leaks instead of explicit step to check for Tor leaks.
Since we merged test/7821-tor into experimental, network capture was
removed as an explicit step in some scenarios in favour of the
@check_tor_leaks tag. Other branches has added scenarios that still
use that step, but since none were started due to the step being
removed from the background they fail.

This is a temporary fix for the experimental branch only.