SSH client: don't proxy connections to 172.17.* to 172.31.*.
The RFC1918 172.16.0.0 range is a /12 one, while our previous configuration
acted as if it was a /16 one, which left aside most of the addresses it should
have covered.

We've fixed that two years ago in the tsocks configuration (commit:35abfc62).
For some reason, we've missed the same issue in the SSH client configuration
back then.

Will-fix: #6558