Retry sending the command in wait_until_remote_shell_is_up.
It seems something in this branch (test/7821-tor) has made this
function fail somethimes, possibly commit 1a55905 (Create the
environment for remote shell calls more realisitcally).

Trying the initial test command several times like in this commit
seems to fix it, and the the client and server in sync. Let's see how
it goes.