Remove OpenPGP verification instructions for the time being (#10675)
Now that we force people to always go through a verification that is as
strong as HTTPS (Browser extension or BitTorrent), the OpenPGP
verification only makes sense for people building trust in our signing
key.

The previous instructions were not taking this into account and should
be rewritten to explain better the thread model and strategy behind a
successful OpenPGP verification, and not so much about the step-by-step
instructions on how to perform this blindly.

But we're running out of time on the installation assistant, so let's
consider that the OpenPGP verification is for people who know how to do
this already and don't need specific instructions.

At least for the time being... This should later on be fixed in #10675.