Make file creation waiting less racy.
I fear (but admittedly haven't seen) that there's a potential race
here; if the [ -e ] check fails, and the file is created before
inotifywait is running, then we will wait endlessly.

We now actually verify that it was the creation of the hostname file
that inotifywait say, and not some other file, like the onion key,
which otherwise would open up for yet another race condition.