Retry updating APT from Synaptic on network errors.
Do not rely on the Synaptic interface to test if the update finished,
given the lack of a clear notification, but test if APT processes are
running, restart if its too long.

Also brings better catch of Synaptic error window, as well as more robust
handling of Synaptic segfaults (if any) by restarting its process.

Refs: #10441