Cherry-pick commit 1bf50e3ccbf68a262d35b8472375ad9fdc29e7fc:
Unmark all scenarios that use Tails Installer as fragile.

bugfix/11590-installer-robustness was based on this branch
(bugfix/10720-installer-freezes-on-jenkins), and flagged the tests as
fragile again, so we have to revert that on this branch so these tests
run on Jenkins again.

This branch is only meant to run more tests in Jenkins involving the
boot menu for Refs: #10777 anyway, it won't be merged.

This reverts the following commits:

  3c02ea74788cffe0b4db21181f46778eab0689b5
  20d07239a8e2c3706827edcf3b17266899c37a64
  2c4c6434c0c6bdcf2534f966434bf92230af3698